Change input format

+++ src/utils/vecs.h
@@ -116,7 +116,7 @@ public:
     long long get_total_msg()
     {
         long long sum = 0;
-        for (int i = 0; i < vecs.size(); i++) {
+        for (size_t i = 0; i < vecs.size(); i++) {
             sum += vecs[i].size();
         }
         return sum;
